// frontend/src/core/models/contactsCstm.ts
/**
 * Created by @ES Express Systems
 * User: Rafael Gutierrez Gaspar
 * Date: Mon Jan 25 2021 02:42:19 GMT-0400 (Bolivia Time)
 * Time: 2:42:19
 * Last User updated: Rafael Gutierrez Gaspar
 * Last date updated: Mon Jan 25 2021 02:42:19 GMT-0400 (Bolivia Time)
 * Last time updated: 2:42:19
 *
 * Caution: es-sections will be replaced by script execution
 */

//<es-section>

//</es-section>

export class ContactsCstm {

  //<es-section>
  
  id_c: string;
  
  
  
  
  
  jjwg_maps_geocode_status_c: string;
  
  jjwg_maps_address_c: string;
  
  
  
  
  
  
  
  jjwg_maps_lng_c: number;
  
  jjwg_maps_lat_c: number;
  
  
  
  
  
  

  //</es-section>
}