// frontend/src/core/models/contactsAudit.ts
/**
 * Created by @ES Express Systems
 * User: Rafael Gutierrez Gaspar
 * Date: Mon Jan 25 2021 02:42:17 GMT-0400 (Bolivia Time)
 * Time: 2:42:17
 * Last User updated: Rafael Gutierrez Gaspar
 * Last date updated: Mon Jan 25 2021 02:42:17 GMT-0400 (Bolivia Time)
 * Last time updated: 2:42:17
 *
 * Caution: es-sections will be replaced by script execution
 */

//<es-section>

//</es-section>

export class ContactsAudit {

  //<es-section>
  
  id: string;
  
  
  
  
  
  created_by: string;
  
  field_name: string;
  
  data_type: string;
  
  before_value_string: string;
  
  after_value_string: string;
  
  
  before_value_text: string;
  
  after_value_text: string;
  
  
  
  date_created: Date;
  
  
  
  parent_id: string;
  
  
  
  
  
  
  

  //</es-section>
}